PCSMOD-48: Moved the PCS Lab rest API from Old rest module to PCS Module
This is done to eliminate dependency of PCS module on the old Rest Module. This is desired
because the later is no longer actively developed meaning that moving forward it could be a
big obstacle whenever upgrading (as it has been the case with 1.9 to 1.11 upgrade).

The reason for native implementation of rest API in PCS is the fact that the existing code
that rely on this API is maintained by someone else and hence we can not risk breaking the
expected behavior

Initially this was implemented as a pass thru to depricated REST Module.
All classes required to implement the minimal features allowing to fetch patients
from AMRS have been moved to this module. This is done solely to maintain the same
interface expected by the PCS Lab System. Otherwise it would have been ideal to
tell the PCS Lab System developers to move their code to use the official OpenMRS
REST webservices module.