AU-11: Fixed a bug where a general ORU-R01 was thrown out during processing.
LabORUR01Handler overrode canProcess method which was implemented to only return true for messages
coming from lab. However the Lab processor falls back to ORUR01 general handler if the message is not from Lab.
This was not happening because when the service calls canProcess(), the subclass version was being called and
it returned false. The method has been renamed to isLabMessage and defer the test to super class.